started to change structure
