+++ composer.json
@@ -1,0 +1,21 @@
+{
+  "name": "users-bulk-delete-with-preview/users-bulk-delete-with-preview",
+  "description": "A WordPress plugin for bulk deleting users and posts with a preview mechanism.",
+  "type": "wordpress-plugin",
+  "license": "GPL-2.0-or-later",
+  "require": {
+    "php": ">=7.4"
+  },
+  "autoload": {
+    "psr-4": {
+      "UsersBulkDeleteWithPreview\\": "src/"
+    }
+  },
+  "authors": [
+    {
+      "name": "Maksym Vasylchuk",
+      "email": "maksymvasylchuk@gmail.com"
+    }
+  ],
+  "minimum-stability": "stable"
+}
